requirements.txt
altgraph==0.17
atomicwrites==1.4.0
attrs==20.3.0
auto-py-to-exe==2.8.0
bleach==3.3.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
docutils==0.16
Eel==0.12.4
entry-points-txt==0.1.0
Flask==1.1.2
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gui2py==0.9.6.1
headerparser==0.4.0
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
MouseInfo==0.1.3
netifaces==0.10.9
numpy==1.20.1
packaging==20.9
pefile==2019.4.18
Pillow==8.1.2
pip-autoremove==0.9.1
pip-gui-tools==0.0.10
plib3.stdlib==1.1.2.post1
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.8.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.5
pytest==6.2.2
PyTweening==1.0.3
pywin32==300
pywin32-ctypes==0.2.0
readme-renderer==24.0
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
wheel-filename==1.2.0
wheel-inspect==1.7.0
whichcraft==0.6.1
WMI==1.5.1
wxPython==4.1.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.2.0
